refactor(extensions): extract disabled adapter
Move the disabled extension adapter implementation out of the normal adapter lifecycle so extension-adapter.ts stays focused on the active runtime path.
